move createTestProject to util
